% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{elem_wise}
\alias{elem_wise}
\title{Element-by-element operations on matrices}
\usage{
env_fwd$elem_wise(f)

env_bwd$elem_wise(f)
}
\description{
Takes an operator which can be applied element-by-element
and converts it into forward and backward functions. The forward function
is trivial (it just returns the original operator). The backward function
uses \link[Deriv:Deriv]{Deriv::Deriv} to automatically compute the derivative (for a unary operator)
or partial derivatives (for a binary operator).
}
\details{
Note that elem_wise counts the number of arguments in the supplied operator.
This can cause surprise errors. For example, \code{log} in base R actually has two arguments
(x and base) so \code{elem_wise} would treat it as a binary operator. To make it unary,
you'd want to define your own single-argument function \code{function (x) log(x)}.
}
\examples{
X = matrix(runif(11*5),nrow=11,ncol=5)
Y = matrix(rnorm(11*5),nrow=11,ncol=5)
prev = matrix(rnorm(11*5),nrow=11,ncol=5)

# unary operator example: exp
env_fwd$elem_wise(exp)(X)
env_bwd$elem_wise(exp)(X,prev)

# unary operator example: log
f = function (x) log(x)
env_fwd$elem_wise(f)(X)
env_bwd$elem_wise(f)(X,prev)

# binary operator example: multiplying matrices element by element
env_fwd$elem_wise(`*`)(X,Y)
env_bwd$elem_wise(`*`)(X,Y,prev)

}
