% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{Operators}
\alias{Operators}
\title{Operators with pre-defined forward and backward functions.}
\description{
Below is listed the operators with pre-defined forward and backward
functions contained in \code{env_fwd} and \code{env_bwd}, respectively. See also \link[=elem_wise]{elem_wise}
for how to deal with operators (either unary or binary) which are applied element
by element to an input matrix.
}
\details{
base R operators:
\itemize{
\item \code{\%*\%}
\item \code{crossprod}
\item \code{tcrossprod}
\item \code{colMeans}
}

operators defined in this package:
\itemize{
\item \link[=loss_L2]{L2 Loss}
\item \link[=softmax]{softmax}
\item \link[=add_colrep]{add matrix to column-repeated vector}
\item \link[=add_rowrep]{add matrix to row-repeated vector}
}
}
